chore: reorganize navigation folders and types (#313)
* chore:renamed screen groups to stack

* chore:create tab folder

* chore: update HomeTabsList location

* chore: update HomeTabsList location

* chore: put tabs in its own file

* chore: consolidate providers

* chore: move app navigator to root

* chore:consolidated shared types

* chore: consolidate typings

* chore: consolidated app stack

* chore:pr cleanup

* chore: add suspense query for device info

* Revert "chore: add suspense query for device info"

This reverts commit 55683c9585f25d7df3d0a9c602148a9fc9e4fe70.

* chore: one off suspense query

* chore: organize navigation types (#324)

* chore: removed unnecessary comment for sharedTypes/index

* chore: copy nav screen types into types/navigation

* chore: moved custom navigation types into sharedTypes/navigation

* chore: update app stack list import

* chore: update DeviceNamingScreen import

* chore: update AppList import

* chore: update HomeTabsList import

* chore: removed InivteProps from sharedType/index

* chore: removed DeviceRole from sharedType/index

* chore: update DeviceRole import

* chore: update DeviceType import

* chore: update NativeHomeTabsNavigationProps imports

* chore: removed unused import

* chore: update NativeNavigationComponent import

* chore: update NativeRootNavigationProps import

* chore: update TabName import

* chore: updates TabBarIconProps import

* chore: device naming sceens to deviceNamingParamsList

* chore: deviceNamingParams

* chore: homeTabsList to HomeTabsParamsList

* chore: update deviceType import

* chore: update device role imports

* chore: plural of params

* chore: update AppStackList to AppStackParamsList

* chore: update env export

* chore: update app navigator with expo splash screen
