+++ src/frontend/App.tsx
@@ -1,28 +1,21 @@
 import * as React from 'react';
-import {useNavigationContainerRef} from '@react-navigation/native';
+import {
+  NavigationContainer,
+  useNavigationContainerRef,
+} from '@react-navigation/native';
 import {createMapeoClient} from '@mapeo/ipc';
-
-import {AppNavigator} from './Navigation/AppNavigator';
-import {AppStackList} from './Navigation/AppStack';
+import {AppNavigator} from './AppNavigator';
 import {IntlProvider} from './contexts/IntlContext';
-import {ApiProvider} from './contexts/ApiContext';
-import {PhotoPromiseProvider} from './contexts/PhotoPromiseContext';
-import {SecurityProvider} from './contexts/SecurityContext';
-import {QueryClient} from '@tanstack/react-query';
 import {MessagePortLike} from './lib/MessagePortLike';
-import {ServerLoading} from './ServerLoading';
-import {ActiveProjectProvider} from './contexts/ProjectContext';
 import {initializeNodejs} from './initializeNodejs';
 import {PermissionsAndroid} from 'react-native';
-import {ExternalProviders} from './contexts/ExternalProviders';
-import {
-  LocalDiscoveryProvider,
-  createLocalDiscoveryController,
-} from './contexts/LocalDiscoveryContext';
+import {AppProviders} from './contexts/AppProviders';
+import {createLocalDiscoveryController} from './contexts/LocalDiscoveryContext';
 import {Loading} from './sharedComponents/Loading';
+import {BottomSheetModalProvider} from '@gorhom/bottom-sheet';
+import {AppStackParamsList} from './sharedTypes/navigation';
 import * as SplashScreen from 'expo-splash-screen';
 
-const queryClient = new QueryClient();
 const messagePort = new MessagePortLike();
 const mapeoApi = createMapeoClient(messagePort, {timeout: Infinity});
 const localDiscoveryController = createLocalDiscoveryController(mapeoApi);
@@ -31,7 +24,7 @@ initializeNodejs();
 SplashScreen.preventAutoHideAsync();
 
 const App = () => {
-  const navRef = useNavigationContainerRef<AppStackList>();
+  const navRef = useNavigationContainerRef<AppStackParamsList>();
   const [permissionsAsked, setPermissionsAsked] = React.useState(false);
   React.useEffect(() => {
     PermissionsAndroid.requestMultiple([
@@ -43,23 +36,18 @@ const App = () => {
 
   return (
     <IntlProvider>
-      <ExternalProviders queryClient={queryClient} navRef={navRef}>
-        <ServerLoading messagePort={messagePort}>
-          <LocalDiscoveryProvider value={localDiscoveryController}>
-            <ApiProvider api={mapeoApi}>
-              <ActiveProjectProvider>
-                <PhotoPromiseProvider>
-                  <SecurityProvider>
-                    <React.Suspense fallback={<Loading />}>
-                      <AppNavigator permissionAsked={permissionsAsked} />
-                    </React.Suspense>
-                  </SecurityProvider>
-                </PhotoPromiseProvider>
-              </ActiveProjectProvider>
-            </ApiProvider>
-          </LocalDiscoveryProvider>
-        </ServerLoading>
-      </ExternalProviders>
+      <AppProviders
+        messagePort={messagePort}
+        localDiscoveryController={localDiscoveryController}
+        mapeoApi={mapeoApi}>
+        <React.Suspense fallback={<Loading />}>
+          <NavigationContainer ref={navRef}>
+            <BottomSheetModalProvider>
+              <AppNavigator permissionAsked={permissionsAsked} />
+            </BottomSheetModalProvider>
+          </NavigationContainer>
+        </React.Suspense>
+      </AppProviders>
     </IntlProvider>
   );
 };
