+++ src/frontend/Navigation/Stack/AppScreens.tsx
@@ -1,14 +1,6 @@
-import {
-  BottomTabNavigationProp,
-  createBottomTabNavigator,
-} from '@react-navigation/bottom-tabs';
-import {NavigatorScreenParams} from '@react-navigation/native';
 import * as React from 'react';
-import {HomeHeader} from '../../sharedComponents/HomeHeader';
-import {RootStack} from '../AppStack';
-import {MessageDescriptor, useIntl} from 'react-intl';
-import {MapScreen} from '../../screens/MapScreen';
-import {CameraScreen} from '../../screens/CameraScreen';
+import {RootStack} from '.';
+import {MessageDescriptor} from 'react-intl';
 import {ObservationEdit} from '../../screens/ObservationEdit';
 import {AddPhotoScreen} from '../../screens/AddPhoto';
 import {AppPasscode} from '../../screens/AppPasscode';
@@ -20,10 +12,6 @@ import {AuthScreen} from '../../screens/AuthScreen';
 import {ObscurePasscode} from '../../screens/ObscurePasscode';
 import {Settings} from '../../screens/Settings';
 import {PresetChooser} from '../../screens/PresetChooser';
-import {
-  createNavigationOptions as CreateObservationsListNavOptions,
-  ObservationsList,
-} from '../../screens/ObservationsList';
 import {ObservationScreen} from '../../screens/Observation';
 import {AppSettings} from '../../screens/Settings/AppSettings';
 import {ProjectSettings} from '../../screens/Settings/ProjectSettings';
@@ -36,7 +24,6 @@ import {ProjectCreated} from '../../screens/Settings/CreateOrJoinProject/CreateP
 import {JoinExistingProject} from '../../screens/Settings/CreateOrJoinProject/JoinExistingProject';
 import {YourTeam} from '../../screens/Settings/ProjectSettings/YourTeam';
 import {SelectDevice} from '../../screens/Settings/ProjectSettings/YourTeam/SelectDevice';
-import {DeviceType} from '../../sharedTypes';
 import {SelectInviteeRole} from '../../screens/Settings/ProjectSettings/YourTeam/SelectInviteeRole';
 import {ReviewInvitation} from '../../screens/Settings/ProjectSettings/YourTeam/ReviewAndInvite/ReviewInvitation';
 import {InviteAccepted} from '../../screens/Settings/ProjectSettings/YourTeam/InviteAccepted';
@@ -53,12 +40,6 @@ import {
   LocationInfoScreen,
   createNavigationOptions as createLocationInfoNavOptions,
 } from '../../screens/LocationInfoScreen';
-import {useCurrentTab} from '../../hooks/useCurrentTab';
-import {TrackingTabBarIcon} from './TabBar/TrackingTabBarIcon';
-import {InviteProps} from '../types';
-import {CameraTabBarIcon} from './TabBar/CameraTabBarIcon';
-import {MapTabBarIcon} from './TabBar/MapTabBarIcon';
-import {SaveTrackScreen} from '../../screens/SaveTrack/SaveTrackScreen';
 import {InviteDeclined} from '../../screens/Settings/ProjectSettings/YourTeam/InviteDeclined';
 import {UnableToCancelInvite} from '../../screens/Settings/ProjectSettings/YourTeam/ReviewAndInvite/UnableToCancelInvite';
 import {SharedLocationContextProvider} from '../../contexts/SharedLocationContext';
@@ -70,144 +51,10 @@ import {
   ManualGpsScreen,
   createNavigationOptions as createManualGpsNavigationOptions,
 } from '../../screens/ManualGpsScreen';
+import {HomeTabs} from '../Tab';
+import {SaveTrackScreen} from '../../screens/SaveTrack/SaveTrackScreen';
 
 export const TAB_BAR_HEIGHT = 70;
-
-export type HomeTabsList = {
-  Map: undefined;
-  Camera: undefined;
-  Tracking: undefined;
-  ObservationsList: undefined;
-};
-
-export type AppList = {
-  Home: NavigatorScreenParams<HomeTabsList>;
-  GpsModal: undefined;
-  Settings: undefined;
-  ProjectConfig: undefined;
-  AboutMapeo: undefined;
-  LanguageSettings: undefined;
-  CoordinateFormat: undefined;
-  Experiments: undefined;
-  PhotosModal: {
-    photoIndex: number;
-    observationId?: string;
-    editing: boolean;
-  };
-  PhotoView: undefined;
-  PresetChooser: undefined;
-  AddPhoto: undefined;
-  Observation: {observationId: string};
-  ObservationEdit: {observationId?: string} | undefined;
-  ManualGpsScreen: undefined;
-  ObservationDetails: {question: number};
-  LeaveProjectScreen: undefined;
-  AlreadyOnProj: undefined;
-  AddToProjectScreen: undefined;
-  UnableToLinkScreen: undefined;
-  ConnectingToDeviceScreen: {task: () => Promise<void>};
-  ConfirmLeavePracticeModeScreen: {projectAction: 'join' | 'create'};
-  CreateProject: undefined;
-  Security: undefined;
-  DirectionalArrow: undefined;
-  P2pUpgrade: undefined;
-  MapSettings: undefined;
-  BackgroundMaps: undefined;
-  BackgroundMapInfo: {
-    bytesStored: number;
-    id: string;
-    styleUrl: string;
-    name: string;
-  };
-  BGMapsSettings: undefined;
-  AuthScreen: undefined;
-  AppPasscode: undefined;
-  ObscurePasscode: undefined;
-  ConfirmPasscodeSheet: {passcode: string};
-  DisablePasscode: undefined;
-  SetPasscode: undefined;
-  EnterPassToTurnOff: undefined;
-  AppSettings: undefined;
-  ProjectSettings: undefined;
-  CreateOrJoinProject: undefined;
-  ProjectCreated: {name: string};
-  JoinExistingProject: undefined;
-  YourTeam: undefined;
-  SelectDevice: undefined;
-  SelectInviteeRole: {name: string; deviceType: DeviceType; deviceId: string};
-  ReviewAndInvite: InviteProps;
-  InviteAccepted: InviteProps;
-  InviteDeclined: InviteProps;
-  UnableToCancelInvite: InviteProps;
-  DeviceNameDisplay: undefined;
-  DeviceNameEdit: undefined;
-  SaveTrack: undefined;
-  Sync: undefined;
-};
-
-const Tab = createBottomTabNavigator<HomeTabsList>();
-
-const HomeTabs = () => {
-  const {handleTabPress} = useCurrentTab();
-  const {formatMessage} = useIntl();
-  return (
-    <Tab.Navigator
-      screenListeners={{
-        tabPress: handleTabPress,
-      }}
-      screenOptions={({route}) => ({
-        tabBarStyle: {height: TAB_BAR_HEIGHT},
-        tabBarShowLabel: false,
-        headerTransparent: true,
-        tabBarTestID: 'tabBarButton' + route.name,
-      })}
-      initialRouteName={'Map'}
-      backBehavior="initialRoute">
-      <Tab.Screen
-        name="ObservationsList"
-        component={ObservationsList}
-        options={CreateObservationsListNavOptions(formatMessage)}
-      />
-      <Tab.Screen
-        name="Map"
-        component={MapScreen}
-        options={{
-          tabBarIcon: MapTabBarIcon,
-          header: HomeHeader,
-        }}
-      />
-      <Tab.Screen
-        name="Camera"
-        component={CameraScreen}
-        options={{
-          tabBarIcon: CameraTabBarIcon,
-          header: HomeHeader,
-        }}
-      />
-
-      {process.env.EXPO_PUBLIC_FEATURE_TRACKS && (
-        <Tab.Screen
-          name="Tracking"
-          options={{
-            tabBarIcon: TrackingTabBarIcon,
-            headerShown: false,
-          }}
-          listeners={({
-            navigation,
-          }: {
-            navigation: BottomTabNavigationProp<HomeTabsList>;
-          }) => ({
-            tabPress: e => {
-              e.preventDefault();
-              navigation.navigate('Map');
-            },
-          })}
-          children={() => <></>}
-        />
-      )}
-    </Tab.Navigator>
-  );
-};
 
 // **NOTE**: No hooks allowed here (this is not a component, it is a function
 // that returns a react element)
