+++ src/frontend/Navigation/Stack/DeviceNamingScreens.tsx
@@ -1,14 +1,8 @@
 import * as React from 'react';
-import {RootStack} from '../AppStack';
+import {RootStack} from '.';
 import {IntroToCoMapeo} from '../../screens/IntroToCoMapeo';
 import {DeviceNaming} from '../../screens/DeviceNaming';
 import {Success} from '../../screens/Success';
-
-export type DeviceNamingSceens = {
-  IntroToCoMapeo: undefined;
-  DeviceNaming: undefined;
-  Success: {deviceName: string};
-};
 
 export const createDeviceNamingScreens = () => (
   <RootStack.Group key="deviceNaming">
