+++ src/frontend/Navigation/Stack/index.tsx
@@ -1,17 +1,11 @@
 import React from 'react';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import {NativeStackNavigationOptions} from '@react-navigation/native-stack/lib/typescript/src/types';
-import {WHITE} from '../lib/styles';
-import {AppList} from './ScreenGroups/AppScreens';
-import {CustomHeaderLeft} from '../sharedComponents/CustomHeaderLeft';
-import {DeviceNamingSceens} from './ScreenGroups/DeviceNamingScreens';
+import {WHITE} from '../../lib/styles';
+import {CustomHeaderLeft} from '../../sharedComponents/CustomHeaderLeft';
+import {AppStackParamsList} from '../../sharedTypes/navigation';
 
-export {createDefaultScreenGroup} from './ScreenGroups/AppScreens';
-
-// export type AppStackList = AppList & OnboardingStackList & IccaStackList;
-export type AppStackList = AppList & DeviceNamingSceens;
-
-export const RootStack = createNativeStackNavigator<AppStackList>();
+export const RootStack = createNativeStackNavigator<AppStackParamsList>();
 
 export const NavigatorScreenOptions: NativeStackNavigationOptions = {
   presentation: 'card',
