+++ src/frontend/Navigation/Tab/TabBar/TabBarIcon.tsx
@@ -1,6 +1,6 @@
 import React, {FC} from 'react';
 import MaterialIcons from 'react-native-vector-icons/MaterialIcons';
-import {TabBarIconProps} from '../../types';
+import {TabBarIconProps} from '../../../sharedTypes/navigation';
 import {COMAPEO_BLUE, MEDIUM_GREY} from '../../../lib/styles';
 
 export interface TabBarIcon extends TabBarIconProps {
