+++ src/frontend/constants.ts
@@ -1,12 +1,11 @@
-import {AppStackList} from './Navigation/AppStack';
-import {HomeTabsList} from './Navigation/ScreenGroups/AppScreens';
+import {AppStackParamsList, HomeTabsParamsList} from './sharedTypes/navigation';
 
 // this has to be a string because js does not recognize 00000 as being 5 digits
 export const OBSCURE_PASSCODE = '00000';
 
 export const EDITING_SCREEN_NAMES:
-  | Omit<keyof AppStackList, 'Home'>[]
-  | (keyof HomeTabsList)[] = [
+  | Omit<keyof AppStackParamsList, 'Home'>[]
+  | (keyof HomeTabsParamsList)[] = [
   'AddPhoto',
   'PresetChooser',
   'ManualGpsScreen',
